backend/requirements.txt
# FastAPI and Server
fastapi==0.110.0
uvicorn[standard]==0.27.0
sse-starlette==2.0.0
python-multipart==0.0.9

# LangChain & LangGraph
langgraph==0.2.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-community==0.3.0

# Vector Database
chromadb==0.5.0

# Database
sqlalchemy==2.0.0
alembic==1.13.0
aiosqlite==0.20.0

# Resume Parsing
python-docx==1.1.0

# Search
duckduckgo-search==3.9.0

# Utils
pydantic==2.6.0
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==23.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
